requirements.txt
Flask==2.3.3
python-multipart==0.0.6
Werkzeug==2.3.7
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
reportlab==4.0.4
